Pulled junit setup for functional tests into a base case and used spring profile to load urls for external apps.
