+++ charts/wa-task-configuration-api/Chart.yaml
@@ -3,7 +3,7 @@ appVersion: "1.0"
 description: A Helm chart for wa-task-configuration-api App
 name: wa-task-configuration-api
 home: https://github.com/hmcts/wa-task-configuration-api
-version: 0.0.5
+version: 0.0.3
 maintainers:
   - name: HMCTS wa team
 dependencies:
