+++ charts/wa-task-configuration-api/values.yaml
@@ -8,7 +8,6 @@ java:
     IDAM_URL: "https://idam-web-public.{{ .Values.global.environment }}.platform.hmcts.net"
     S2S_URL: "http://rpe-service-auth-provider-{{ .Values.global.environment }}.service.core-compute-{{ .Values.global.environment }}.internal"
     CCD_URL: "http://ccd-data-store-api-{{ .Values.global.environment }}.service.core-compute-{{ .Values.global.environment }}.internal"
-    FORCE_REBUILD: 1
   keyVaults:
     wa:
       secrets:
