+++ src/functionalTest/java/uk/gov/hmcts/reform/wataskconfigurationapi/BaseFunctionalTest.java
@@ -1,0 +1,31 @@
+package uk.gov.hmcts.reform.wataskconfigurationapi;
+
+import io.restassured.RestAssured;
+import net.serenitybdd.junit.spring.integration.SpringIntegrationSerenityRunner;
+import org.junit.Before;
+import org.junit.runner.RunWith;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.test.context.ActiveProfiles;
+
+import java.time.format.DateTimeFormatter;
+
+import static java.time.format.DateTimeFormatter.ofPattern;
+
+@RunWith(SpringIntegrationSerenityRunner.class)
+@SpringBootTest
+@ActiveProfiles("functional")
+public abstract class BaseFunctionalTest {
+    public static final DateTimeFormatter CAMUNDA_DATA_TIME_FORMATTER = ofPattern("yyyy-MM-dd'T'HH:mm:ss.SSSZ");
+
+    @Value("${targets.instance}")
+    protected String testUrl;
+    @Value("${targets.camunda}")
+    protected String camundaUrl;
+
+    @Before
+    public void setUp() {
+        RestAssured.baseURI = testUrl;
+        RestAssured.useRelaxedHTTPSValidation();
+    }
+}
