+++ src/functionalTest/java/uk/gov/hmcts/reform/wataskconfigurationapi/WelcomeTest.java
@@ -1,31 +1,13 @@
 package uk.gov.hmcts.reform.wataskconfigurationapi;
 
-import io.restassured.RestAssured;
-import net.serenitybdd.junit.spring.integration.SpringIntegrationSerenityRunner;
 import net.serenitybdd.rest.SerenityRest;
-import org.junit.Before;
 import org.junit.Test;
-import org.junit.runner.RunWith;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
-import org.springframework.test.context.ActiveProfiles;
 
 import static org.hamcrest.Matchers.containsString;
 
-@RunWith(SpringIntegrationSerenityRunner.class)
-@SpringBootTest
-@ActiveProfiles("functional")
-public class WelcomeTest {
-
-    @Value("${targetInstance}") private String testUrl;
-
-    @Before
-    public void setUp() {
-        RestAssured.baseURI = testUrl;
-        RestAssured.useRelaxedHTTPSValidation();
-    }
+public class WelcomeTest extends BaseFunctionalTest {
 
     @Test
     public void should_welcome_with_200_response_code() {
