+++ src/functionalTest/resources/application-functional.yaml
@@ -1,1 +1,3 @@
-targetInstance: ${TEST_URL:http://localhost:8091}
+targets:
+  instance: ${TEST_URL:http://localhost:8091}
+  camunda: ${CAMUNDA_URL:http://localhost:8080/engine-rest}
